Introduce empty state for Pool Score list and "Create Pool" action
This commit adds an empty state to the Pool Score list in the iOS app. When a user has no pools, they will now see a message "Your pools are waiting" with a subtitle "Be the first to create one and invite others to join", along with a "Create" button.